+++ src/main/java/team7/inplace/crawling/client/YoutubeClient.java
@@ -10,7 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
-import team7.inplace.crawling.application.dto.RawVideoInfo;
+import team7.inplace.crawling.client.dto.RawVideoInfo;
 
 @Slf4j
 @Component
