+++ src/test/java/team7/inplace/crawling/client/KakaoMapClientTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import team7.inplace.crawling.application.dto.RawVideoInfo;
+import team7.inplace.crawling.client.dto.RawVideoInfo;
 
 @SpringBootTest
 @ActiveProfiles("test")
